# igor_homework_api_specification.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8090","description":"Generated server url"}],"paths":{"/v1/employees":{"get":{"tags":["employee-controller"],"operationId":"getEmployees","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"team","in":"query","required":false,"schema":{"type":"string"}},{"name":"teamLead","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Employee"}}}}}}},"post":{"tags":["employee-controller"],"operationId":"addEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}}},"/v1/employees/{employeeId}":{"get":{"tags":["employee-controller"],"operationId":"getEmployeeById","parameters":[{"name":"employeeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}},"delete":{"tags":["employee-controller"],"operationId":"deleteEmployee","parameters":[{"name":"employeeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Error":{"type":"object","properties":{"httpCode":{"type":"integer","format":"int32"},"message":{"type":"string"}}},"Employee":{"type":"object","properties":{"personalId":{"type":"integer","format":"int64"},"name":{"type":"string"},"team":{"type":"string"},"teamLead":{"type":"string"}}}}}}